分离依赖
