+++ commons.py
@@ -1,7 +1,6 @@
 import math
-import numpy as np
+
 import torch
-from torch import nn
 from torch.nn import functional as F
 
 
@@ -12,7 +11,7 @@ def init_weights(m, mean=0.0, std=0.01):
 
 
 def get_padding(kernel_size, dilation=1):
-  return int((kernel_size*dilation - dilation)/2)
+  return int((kernel_size * dilation - dilation) / 2)
 
 
 def convert_pad_shape(pad_shape):
@@ -30,7 +29,7 @@ def intersperse(lst, item):
 def kl_divergence(m_p, logs_p, m_q, logs_q):
   """KL(P||Q)"""
   kl = (logs_q - logs_p) - 0.5
-  kl += 0.5 * (torch.exp(2. * logs_p) + ((m_p - m_q)**2)) * torch.exp(-2. * logs_q)
+  kl += 0.5 * (torch.exp(2. * logs_p) + ((m_p - m_q) ** 2)) * torch.exp(-2. * logs_q)
   return kl
 
 
@@ -72,7 +71,7 @@ def get_timing_signal_1d(
       math.log(float(max_timescale) / float(min_timescale)) /
       (num_timescales - 1))
   inv_timescales = min_timescale * torch.exp(
-      torch.arange(num_timescales, dtype=torch.float) * -log_timescale_increment)
+    torch.arange(num_timescales, dtype=torch.float) * -log_timescale_increment)
   scaled_time = position.unsqueeze(0) * inv_timescales.unsqueeze(1)
   signal = torch.cat([torch.sin(scaled_time), torch.cos(scaled_time)], 0)
   signal = F.pad(signal, [0, 0, 0, channels % 2])
@@ -131,15 +130,15 @@ def generate_path(duration, mask):
   mask: [b, 1, t_y, t_x]
   """
   device = duration.device
-  
+
   b, _, t_y, t_x = mask.shape
   cum_duration = torch.cumsum(duration, -1)
-  
+
   cum_duration_flat = cum_duration.view(b * t_x)
   path = sequence_mask(cum_duration_flat, t_y).to(mask.dtype)
   path = path.view(b, t_x, t_y)
   path = path - F.pad(path, convert_pad_shape([[0, 0], [1, 0], [0, 0]]))[:, :-1]
-  path = path.unsqueeze(1).transpose(2,3) * mask
+  path = path.unsqueeze(1).transpose(2, 3) * mask
   return path
