+++ losses.py
@@ -1,7 +1,4 @@
-import torch 
-from torch.nn import functional as F
-
-import commons
+import torch
 
 
 def feature_loss(fmap_r, fmap_g):
@@ -12,7 +9,7 @@ def feature_loss(fmap_r, fmap_g):
       gl = gl.float()
       loss += torch.mean(torch.abs(rl - gl))
 
-  return loss * 2 
+  return loss * 2
 
 
 def discriminator_loss(disc_real_outputs, disc_generated_outputs):
@@ -22,8 +19,8 @@ def discriminator_loss(disc_real_outputs, disc_generated_outputs):
   for dr, dg in zip(disc_real_outputs, disc_generated_outputs):
     dr = dr.float()
     dg = dg.float()
-    r_loss = torch.mean((1-dr)**2)
-    g_loss = torch.mean(dg**2)
+    r_loss = torch.mean((1 - dr) ** 2)
+    g_loss = torch.mean(dg ** 2)
     loss += (r_loss + g_loss)
     r_losses.append(r_loss.item())
     g_losses.append(g_loss.item())
@@ -36,7 +33,7 @@ def generator_loss(disc_outputs):
   gen_losses = []
   for dg in disc_outputs:
     dg = dg.float()
-    l = torch.mean((1-dg)**2)
+    l = torch.mean((1 - dg) ** 2)
     gen_losses.append(l)
     loss += l
 
@@ -55,7 +52,7 @@ def kl_loss(z_p, logs_q, m_p, logs_p, z_mask):
   z_mask = z_mask.float()
 
   kl = logs_p - logs_q - 0.5
-  kl += 0.5 * ((z_p - m_p)**2) * torch.exp(-2. * logs_p)
+  kl += 0.5 * ((z_p - m_p) ** 2) * torch.exp(-2. * logs_p)
   kl = torch.sum(kl * z_mask)
   l = kl / torch.sum(z_mask)
   return l
