+++ mel_processing.py
@@ -7,8 +7,6 @@ import torch.nn.functional as F
 import torch.utils.data
 import numpy as np
 import librosa
-import librosa.util as librosa_util
-from librosa.util import normalize, pad_center, tiny
 from scipy.signal import get_window
 from scipy.io.wavfile import read
 from librosa.filters import mel as librosa_mel_fn
@@ -17,31 +15,31 @@ MAX_WAV_VALUE = 32768.0
 
 
 def dynamic_range_compression_torch(x, C=1, clip_val=1e-5):
-    """
-    PARAMS
-    ------
-    C: compression factor
-    """
-    return torch.log(torch.clamp(x, min=clip_val) * C)
+  """
+  PARAMS
+  ------
+  C: compression factor
+  """
+  return torch.log(torch.clamp(x, min=clip_val) * C)
 
 
 def dynamic_range_decompression_torch(x, C=1):
-    """
-    PARAMS
-    ------
-    C: compression factor used to compress
-    """
-    return torch.exp(x) / C
+  """
+  PARAMS
+  ------
+  C: compression factor used to compress
+  """
+  return torch.exp(x) / C
 
 
 def spectral_normalize_torch(magnitudes):
-    output = dynamic_range_compression_torch(magnitudes)
-    return output
+  output = dynamic_range_compression_torch(magnitudes)
+  return output
 
 
 def spectral_de_normalize_torch(magnitudes):
-    output = dynamic_range_decompression_torch(magnitudes)
-    return output
+  output = dynamic_range_decompression_torch(magnitudes)
+  return output
 
 
 mel_basis = {}
@@ -49,64 +47,66 @@ hann_window = {}
 
 
 def spectrogram_torch(y, n_fft, sampling_rate, hop_size, win_size, center=False):
-    if torch.min(y) < -1.:
-        print('min value is ', torch.min(y))
-    if torch.max(y) > 1.:
-        print('max value is ', torch.max(y))
+  if torch.min(y) < -1.:
+    print('min value is ', torch.min(y))
+  if torch.max(y) > 1.:
+    print('max value is ', torch.max(y))
 
-    global hann_window
-    dtype_device = str(y.dtype) + '_' + str(y.device)
-    wnsize_dtype_device = str(win_size) + '_' + dtype_device
-    if wnsize_dtype_device not in hann_window:
-        hann_window[wnsize_dtype_device] = torch.hann_window(win_size).to(dtype=y.dtype, device=y.device)
+  global hann_window
+  dtype_device = str(y.dtype) + '_' + str(y.device)
+  wnsize_dtype_device = str(win_size) + '_' + dtype_device
+  if wnsize_dtype_device not in hann_window:
+    hann_window[wnsize_dtype_device] = torch.hann_window(win_size).to(dtype=y.dtype, device=y.device)
 
-    y = torch.nn.functional.pad(y.unsqueeze(1), (int((n_fft-hop_size)/2), int((n_fft-hop_size)/2)), mode='reflect')
-    y = y.squeeze(1)
+  y = torch.nn.functional.pad(y.unsqueeze(1), (
+  int((n_fft - hop_size) / 2), int((n_fft - hop_size) / 2)), mode='reflect')
+  y = y.squeeze(1)
 
-    spec = torch.stft(y, n_fft, hop_length=hop_size, win_length=win_size, window=hann_window[wnsize_dtype_device],
-                      center=center, pad_mode='reflect', normalized=False, onesided=True)
+  spec = torch.stft(y, n_fft, hop_length=hop_size, win_length=win_size, window=hann_window[wnsize_dtype_device],
+                    center=center, pad_mode='reflect', normalized=False, onesided=True)
 
-    spec = torch.sqrt(spec.pow(2).sum(-1) + 1e-6)
-    return spec
+  spec = torch.sqrt(spec.pow(2).sum(-1) + 1e-6)
+  return spec
 
 
 def spec_to_mel_torch(spec, n_fft, num_mels, sampling_rate, fmin, fmax):
-    global mel_basis
-    dtype_device = str(spec.dtype) + '_' + str(spec.device)
-    fmax_dtype_device = str(fmax) + '_' + dtype_device
-    if fmax_dtype_device not in mel_basis:
-        mel = librosa_mel_fn(sampling_rate, n_fft, num_mels, fmin, fmax)
-        mel_basis[fmax_dtype_device] = torch.from_numpy(mel).to(dtype=spec.dtype, device=spec.device)
-    spec = torch.matmul(mel_basis[fmax_dtype_device], spec)
-    spec = spectral_normalize_torch(spec)
-    return spec
+  global mel_basis
+  dtype_device = str(spec.dtype) + '_' + str(spec.device)
+  fmax_dtype_device = str(fmax) + '_' + dtype_device
+  if fmax_dtype_device not in mel_basis:
+    mel = librosa_mel_fn(sampling_rate, n_fft, num_mels, fmin, fmax)
+    mel_basis[fmax_dtype_device] = torch.from_numpy(mel).to(dtype=spec.dtype, device=spec.device)
+  spec = torch.matmul(mel_basis[fmax_dtype_device], spec)
+  spec = spectral_normalize_torch(spec)
+  return spec
 
 
 def mel_spectrogram_torch(y, n_fft, num_mels, sampling_rate, hop_size, win_size, fmin, fmax, center=False):
-    if torch.min(y) < -1.:
-        print('min value is ', torch.min(y))
-    if torch.max(y) > 1.:
-        print('max value is ', torch.max(y))
+  if torch.min(y) < -1.:
+    print('min value is ', torch.min(y))
+  if torch.max(y) > 1.:
+    print('max value is ', torch.max(y))
 
-    global mel_basis, hann_window
-    dtype_device = str(y.dtype) + '_' + str(y.device)
-    fmax_dtype_device = str(fmax) + '_' + dtype_device
-    wnsize_dtype_device = str(win_size) + '_' + dtype_device
-    if fmax_dtype_device not in mel_basis:
-        mel = librosa_mel_fn(sampling_rate, n_fft, num_mels, fmin, fmax)
-        mel_basis[fmax_dtype_device] = torch.from_numpy(mel).to(dtype=y.dtype, device=y.device)
-    if wnsize_dtype_device not in hann_window:
-        hann_window[wnsize_dtype_device] = torch.hann_window(win_size).to(dtype=y.dtype, device=y.device)
+  global mel_basis, hann_window
+  dtype_device = str(y.dtype) + '_' + str(y.device)
+  fmax_dtype_device = str(fmax) + '_' + dtype_device
+  wnsize_dtype_device = str(win_size) + '_' + dtype_device
+  if fmax_dtype_device not in mel_basis:
+    mel = librosa_mel_fn(sampling_rate, n_fft, num_mels, fmin, fmax)
+    mel_basis[fmax_dtype_device] = torch.from_numpy(mel).to(dtype=y.dtype, device=y.device)
+  if wnsize_dtype_device not in hann_window:
+    hann_window[wnsize_dtype_device] = torch.hann_window(win_size).to(dtype=y.dtype, device=y.device)
 
-    y = torch.nn.functional.pad(y.unsqueeze(1), (int((n_fft-hop_size)/2), int((n_fft-hop_size)/2)), mode='reflect')
-    y = y.squeeze(1)
+  y = torch.nn.functional.pad(y.unsqueeze(1), (
+  int((n_fft - hop_size) / 2), int((n_fft - hop_size) / 2)), mode='reflect')
+  y = y.squeeze(1)
 
-    spec = torch.stft(y, n_fft, hop_length=hop_size, win_length=win_size, window=hann_window[wnsize_dtype_device],
-                      center=center, pad_mode='reflect', normalized=False, onesided=True)
+  spec = torch.stft(y, n_fft, hop_length=hop_size, win_length=win_size, window=hann_window[wnsize_dtype_device],
+                    center=center, pad_mode='reflect', normalized=False, onesided=True)
 
-    spec = torch.sqrt(spec.pow(2).sum(-1) + 1e-6)
+  spec = torch.sqrt(spec.pow(2).sum(-1) + 1e-6)
 
-    spec = torch.matmul(mel_basis[fmax_dtype_device], spec)
-    spec = spectral_normalize_torch(spec)
+  spec = torch.matmul(mel_basis[fmax_dtype_device], spec)
+  spec = spectral_normalize_torch(spec)
 
-    return spec
+  return spec
