+++ modules.py
@@ -1,18 +1,14 @@
-import copy
 import math
-import numpy as np
-import scipy
+
 import torch
 from torch import nn
+from torch.nn import Conv1d
 from torch.nn import functional as F
-
-from torch.nn import Conv1d, ConvTranspose1d, AvgPool1d, Conv2d
 from torch.nn.utils import weight_norm, remove_weight_norm
 
 import commons
 from commons import init_weights, get_padding
 from transforms import piecewise_rational_quadratic_transform
-
 
 LRELU_SLOPE = 0.1
 
@@ -31,7 +27,7 @@ class LayerNorm(nn.Module):
     x = F.layer_norm(x, (self.channels,), self.gamma, self.beta, self.eps)
     return x.transpose(1, -1)
 
- 
+
 class ConvReluNorm(nn.Module):
   def __init__(self, in_channels, hidden_channels, out_channels, kernel_size, n_layers, p_dropout):
     super().__init__()
@@ -45,13 +41,13 @@ class ConvReluNorm(nn.Module):
 
     self.conv_layers = nn.ModuleList()
     self.norm_layers = nn.ModuleList()
-    self.conv_layers.append(nn.Conv1d(in_channels, hidden_channels, kernel_size, padding=kernel_size//2))
+    self.conv_layers.append(nn.Conv1d(in_channels, hidden_channels, kernel_size, padding=kernel_size // 2))
     self.norm_layers.append(LayerNorm(hidden_channels))
     self.relu_drop = nn.Sequential(
-        nn.ReLU(),
-        nn.Dropout(p_dropout))
-    for _ in range(n_layers-1):
-      self.conv_layers.append(nn.Conv1d(hidden_channels, hidden_channels, kernel_size, padding=kernel_size//2))
+      nn.ReLU(),
+      nn.Dropout(p_dropout))
+    for _ in range(n_layers - 1):
+      self.conv_layers.append(nn.Conv1d(hidden_channels, hidden_channels, kernel_size, padding=kernel_size // 2))
       self.norm_layers.append(LayerNorm(hidden_channels))
     self.proj = nn.Conv1d(hidden_channels, out_channels, 1)
     self.proj.weight.data.zero_()
@@ -71,6 +67,7 @@ class DDSConv(nn.Module):
   """
   Dialted and Depth-Separable Convolution
   """
+
   def __init__(self, channels, kernel_size, n_layers, p_dropout=0.):
     super().__init__()
     self.channels = channels
@@ -86,9 +83,9 @@ class DDSConv(nn.Module):
     for i in range(n_layers):
       dilation = kernel_size ** i
       padding = (kernel_size * dilation - dilation) // 2
-      self.convs_sep.append(nn.Conv1d(channels, channels, kernel_size, 
-          groups=channels, dilation=dilation, padding=padding
-      ))
+      self.convs_sep.append(nn.Conv1d(channels, channels, kernel_size,
+                                      groups=channels, dilation=dilation, padding=padding
+                                      ))
       self.convs_1x1.append(nn.Conv1d(channels, channels, 1))
       self.norms_1.append(LayerNorm(channels))
       self.norms_2.append(LayerNorm(channels))
@@ -111,8 +108,8 @@ class DDSConv(nn.Module):
 class WN(torch.nn.Module):
   def __init__(self, hidden_channels, kernel_size, dilation_rate, n_layers, gin_channels=0, p_dropout=0):
     super(WN, self).__init__()
-    assert(kernel_size % 2 == 1)
-    self.hidden_channels =hidden_channels
+    assert (kernel_size % 2 == 1)
+    self.hidden_channels = hidden_channels
     self.kernel_size = kernel_size,
     self.dilation_rate = dilation_rate
     self.n_layers = n_layers
@@ -124,13 +121,13 @@ class WN(torch.nn.Module):
     self.drop = nn.Dropout(p_dropout)
 
     if gin_channels != 0:
-      cond_layer = torch.nn.Conv1d(gin_channels, 2*hidden_channels*n_layers, 1)
+      cond_layer = torch.nn.Conv1d(gin_channels, 2 * hidden_channels * n_layers, 1)
       self.cond_layer = torch.nn.utils.weight_norm(cond_layer, name='weight')
 
     for i in range(n_layers):
       dilation = dilation_rate ** i
       padding = int((kernel_size * dilation - dilation) / 2)
-      in_layer = torch.nn.Conv1d(hidden_channels, 2*hidden_channels, kernel_size,
+      in_layer = torch.nn.Conv1d(hidden_channels, 2 * hidden_channels, kernel_size,
                                  dilation=dilation, padding=padding)
       in_layer = torch.nn.utils.weight_norm(in_layer, name='weight')
       self.in_layers.append(in_layer)
@@ -156,21 +153,21 @@ class WN(torch.nn.Module):
       x_in = self.in_layers[i](x)
       if g is not None:
         cond_offset = i * 2 * self.hidden_channels
-        g_l = g[:,cond_offset:cond_offset+2*self.hidden_channels,:]
+        g_l = g[:, cond_offset:cond_offset + 2 * self.hidden_channels, :]
       else:
         g_l = torch.zeros_like(x_in)
 
       acts = commons.fused_add_tanh_sigmoid_multiply(
-          x_in,
-          g_l,
-          n_channels_tensor)
+        x_in,
+        g_l,
+        n_channels_tensor)
       acts = self.drop(acts)
 
       res_skip_acts = self.res_skip_layers[i](acts)
       if i < self.n_layers - 1:
-        res_acts = res_skip_acts[:,:self.hidden_channels,:]
+        res_acts = res_skip_acts[:, :self.hidden_channels, :]
         x = (x + res_acts) * x_mask
-        output = output + res_skip_acts[:,self.hidden_channels:,:]
+        output = output + res_skip_acts[:, self.hidden_channels:, :]
       else:
         output = output + res_skip_acts
     return output * x_mask
@@ -181,79 +178,79 @@ class WN(torch.nn.Module):
     for l in self.in_layers:
       torch.nn.utils.remove_weight_norm(l)
     for l in self.res_skip_layers:
-     torch.nn.utils.remove_weight_norm(l)
+      torch.nn.utils.remove_weight_norm(l)
 
 
 class ResBlock1(torch.nn.Module):
-    def __init__(self, channels, kernel_size=3, dilation=(1, 3, 5)):
-        super(ResBlock1, self).__init__()
-        self.convs1 = nn.ModuleList([
-            weight_norm(Conv1d(channels, channels, kernel_size, 1, dilation=dilation[0],
-                               padding=get_padding(kernel_size, dilation[0]))),
-            weight_norm(Conv1d(channels, channels, kernel_size, 1, dilation=dilation[1],
-                               padding=get_padding(kernel_size, dilation[1]))),
-            weight_norm(Conv1d(channels, channels, kernel_size, 1, dilation=dilation[2],
-                               padding=get_padding(kernel_size, dilation[2])))
-        ])
-        self.convs1.apply(init_weights)
+  def __init__(self, channels, kernel_size=3, dilation=(1, 3, 5)):
+    super(ResBlock1, self).__init__()
+    self.convs1 = nn.ModuleList([
+      weight_norm(Conv1d(channels, channels, kernel_size, 1, dilation=dilation[0],
+                         padding=get_padding(kernel_size, dilation[0]))),
+      weight_norm(Conv1d(channels, channels, kernel_size, 1, dilation=dilation[1],
+                         padding=get_padding(kernel_size, dilation[1]))),
+      weight_norm(Conv1d(channels, channels, kernel_size, 1, dilation=dilation[2],
+                         padding=get_padding(kernel_size, dilation[2])))
+    ])
+    self.convs1.apply(init_weights)
 
-        self.convs2 = nn.ModuleList([
-            weight_norm(Conv1d(channels, channels, kernel_size, 1, dilation=1,
-                               padding=get_padding(kernel_size, 1))),
-            weight_norm(Conv1d(channels, channels, kernel_size, 1, dilation=1,
-                               padding=get_padding(kernel_size, 1))),
-            weight_norm(Conv1d(channels, channels, kernel_size, 1, dilation=1,
-                               padding=get_padding(kernel_size, 1)))
-        ])
-        self.convs2.apply(init_weights)
+    self.convs2 = nn.ModuleList([
+      weight_norm(Conv1d(channels, channels, kernel_size, 1, dilation=1,
+                         padding=get_padding(kernel_size, 1))),
+      weight_norm(Conv1d(channels, channels, kernel_size, 1, dilation=1,
+                         padding=get_padding(kernel_size, 1))),
+      weight_norm(Conv1d(channels, channels, kernel_size, 1, dilation=1,
+                         padding=get_padding(kernel_size, 1)))
+    ])
+    self.convs2.apply(init_weights)
 
-    def forward(self, x, x_mask=None):
-        for c1, c2 in zip(self.convs1, self.convs2):
-            xt = F.leaky_relu(x, LRELU_SLOPE)
-            if x_mask is not None:
-                xt = xt * x_mask
-            xt = c1(xt)
-            xt = F.leaky_relu(xt, LRELU_SLOPE)
-            if x_mask is not None:
-                xt = xt * x_mask
-            xt = c2(xt)
-            x = xt + x
-        if x_mask is not None:
-            x = x * x_mask
-        return x
+  def forward(self, x, x_mask=None):
+    for c1, c2 in zip(self.convs1, self.convs2):
+      xt = F.leaky_relu(x, LRELU_SLOPE)
+      if x_mask is not None:
+        xt = xt * x_mask
+      xt = c1(xt)
+      xt = F.leaky_relu(xt, LRELU_SLOPE)
+      if x_mask is not None:
+        xt = xt * x_mask
+      xt = c2(xt)
+      x = xt + x
+    if x_mask is not None:
+      x = x * x_mask
+    return x
 
-    def remove_weight_norm(self):
-        for l in self.convs1:
-            remove_weight_norm(l)
-        for l in self.convs2:
-            remove_weight_norm(l)
+  def remove_weight_norm(self):
+    for l in self.convs1:
+      remove_weight_norm(l)
+    for l in self.convs2:
+      remove_weight_norm(l)
 
 
 class ResBlock2(torch.nn.Module):
-    def __init__(self, channels, kernel_size=3, dilation=(1, 3)):
-        super(ResBlock2, self).__init__()
-        self.convs = nn.ModuleList([
-            weight_norm(Conv1d(channels, channels, kernel_size, 1, dilation=dilation[0],
-                               padding=get_padding(kernel_size, dilation[0]))),
-            weight_norm(Conv1d(channels, channels, kernel_size, 1, dilation=dilation[1],
-                               padding=get_padding(kernel_size, dilation[1])))
-        ])
-        self.convs.apply(init_weights)
+  def __init__(self, channels, kernel_size=3, dilation=(1, 3)):
+    super(ResBlock2, self).__init__()
+    self.convs = nn.ModuleList([
+      weight_norm(Conv1d(channels, channels, kernel_size, 1, dilation=dilation[0],
+                         padding=get_padding(kernel_size, dilation[0]))),
+      weight_norm(Conv1d(channels, channels, kernel_size, 1, dilation=dilation[1],
+                         padding=get_padding(kernel_size, dilation[1])))
+    ])
+    self.convs.apply(init_weights)
 
-    def forward(self, x, x_mask=None):
-        for c in self.convs:
-            xt = F.leaky_relu(x, LRELU_SLOPE)
-            if x_mask is not None:
-                xt = xt * x_mask
-            xt = c(xt)
-            x = xt + x
-        if x_mask is not None:
-            x = x * x_mask
-        return x
+  def forward(self, x, x_mask=None):
+    for c in self.convs:
+      xt = F.leaky_relu(x, LRELU_SLOPE)
+      if x_mask is not None:
+        xt = xt * x_mask
+      xt = c(xt)
+      x = xt + x
+    if x_mask is not None:
+      x = x * x_mask
+    return x
 
-    def remove_weight_norm(self):
-        for l in self.convs:
-            remove_weight_norm(l)
+  def remove_weight_norm(self):
+    for l in self.convs:
+      remove_weight_norm(l)
 
 
 class Log(nn.Module):
@@ -265,7 +262,7 @@ class Log(nn.Module):
     else:
       x = torch.exp(x) * x_mask
       return x
-    
+
 
 class Flip(nn.Module):
   def forward(self, x, *args, reverse=False, **kwargs):
@@ -281,14 +278,14 @@ class ElementwiseAffine(nn.Module):
   def __init__(self, channels):
     super().__init__()
     self.channels = channels
-    self.m = nn.Parameter(torch.zeros(channels,1))
-    self.logs = nn.Parameter(torch.zeros(channels,1))
+    self.m = nn.Parameter(torch.zeros(channels, 1))
+    self.logs = nn.Parameter(torch.zeros(channels, 1))
 
   def forward(self, x, x_mask, reverse=False, **kwargs):
     if not reverse:
       y = self.m + torch.exp(self.logs) * x
       y = y * x_mask
-      logdet = torch.sum(self.logs * x_mask, [1,2])
+      logdet = torch.sum(self.logs * x_mask, [1, 2])
       return y, logdet
     else:
       x = (x - self.m) * torch.exp(-self.logs) * x_mask
@@ -297,14 +294,14 @@ class ElementwiseAffine(nn.Module):
 
 class ResidualCouplingLayer(nn.Module):
   def __init__(self,
-      channels,
-      hidden_channels,
-      kernel_size,
-      dilation_rate,
-      n_layers,
-      p_dropout=0,
-      gin_channels=0,
-      mean_only=False):
+               channels,
+               hidden_channels,
+               kernel_size,
+               dilation_rate,
+               n_layers,
+               p_dropout=0,
+               gin_channels=0,
+               mean_only=False):
     assert channels % 2 == 0, "channels should be divisible by 2"
     super().__init__()
     self.channels = channels
@@ -322,12 +319,12 @@ class ResidualCouplingLayer(nn.Module):
     self.post.bias.data.zero_()
 
   def forward(self, x, x_mask, g=None, reverse=False):
-    x0, x1 = torch.split(x, [self.half_channels]*2, 1)
+    x0, x1 = torch.split(x, [self.half_channels] * 2, 1)
     h = self.pre(x0) * x_mask
     h = self.enc(h, x_mask, g=g)
     stats = self.post(h) * x_mask
     if not self.mean_only:
-      m, logs = torch.split(stats, [self.half_channels]*2, 1)
+      m, logs = torch.split(stats, [self.half_channels] * 2, 1)
     else:
       m = stats
       logs = torch.zeros_like(m)
@@ -335,7 +332,7 @@ class ResidualCouplingLayer(nn.Module):
     if not reverse:
       x1 = m + x1 * torch.exp(logs) * x_mask
       x = torch.cat([x0, x1], 1)
-      logdet = torch.sum(logs, [1,2])
+      logdet = torch.sum(logs, [1, 2])
       return x, logdet
     else:
       x1 = (x1 - m) * torch.exp(-logs) * x_mask
@@ -361,30 +358,30 @@ class ConvFlow(nn.Module):
     self.proj.bias.data.zero_()
 
   def forward(self, x, x_mask, g=None, reverse=False):
-    x0, x1 = torch.split(x, [self.half_channels]*2, 1)
+    x0, x1 = torch.split(x, [self.half_channels] * 2, 1)
     h = self.pre(x0)
     h = self.convs(h, x_mask, g=g)
     h = self.proj(h) * x_mask
 
     b, c, t = x0.shape
-    h = h.reshape(b, c, -1, t).permute(0, 1, 3, 2) # [b, cx?, t] -> [b, c, t, ?]
+    h = h.reshape(b, c, -1, t).permute(0, 1, 3, 2)  # [b, cx?, t] -> [b, c, t, ?]
 
     unnormalized_widths = h[..., :self.num_bins] / math.sqrt(self.filter_channels)
-    unnormalized_heights = h[..., self.num_bins:2*self.num_bins] / math.sqrt(self.filter_channels)
+    unnormalized_heights = h[..., self.num_bins:2 * self.num_bins] / math.sqrt(self.filter_channels)
     unnormalized_derivatives = h[..., 2 * self.num_bins:]
 
     x1, logabsdet = piecewise_rational_quadratic_transform(x1,
-        unnormalized_widths,
-        unnormalized_heights,
-        unnormalized_derivatives,
-        inverse=reverse,
-        tails='linear',
-        tail_bound=self.tail_bound
-    )
+                                                           unnormalized_widths,
+                                                           unnormalized_heights,
+                                                           unnormalized_derivatives,
+                                                           inverse=reverse,
+                                                           tails='linear',
+                                                           tail_bound=self.tail_bound
+                                                           )
 
     x = torch.cat([x0, x1], 1) * x_mask
-    logdet = torch.sum(logabsdet * x_mask, [1,2])
+    logdet = torch.sum(logabsdet * x_mask, [1, 2])
     if not reverse:
-        return x, logdet
+      return x, logdet
     else:
-        return x
+      return x
